Fix: Move appendHoverText mixin to client-only class to prevent dedicated server crash
